Move main() function from "interpreter.c" to "main.c"
